pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "datasette-files"
version = "0.1"
description = "File management for Datasette"
readme = "README.md"
authors = [{name = "Datasette"}]
license = "Apache-2.0"
classifiers=[
    "Framework :: Datasette"
]
requires-python = ">=3.10"
dependencies = [
    "datasette>=1.0a24",
    "python-ulid",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.urls]
Homepage = "https://github.com/datasette/datasette-files"
Changelog = "https://github.com/datasette/datasette-files/releases"
Issues = "https://github.com/datasette/datasette-files/issues"
CI = "https://github.com/datasette/datasette-files/actions"

[project.entry-points.datasette]
files = "datasette_files"

[tool.setuptools.packages.find]
include = ["datasette_files"]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.setuptools.package-data]
datasette_files = ["templates/*", "static/*"]
